# Conflicts:
#	knitkit/app/assets/javascripts/erp_app/desktop/applications/knitkit/splash_screen.js

+++ erp_tech_svcs/lib/erp_tech_svcs/file_support/file_system_manager.rb
@@ -12,7 +12,10 @@ module ErpTechSvcs
         create_path = new_path.split('/').reverse.drop(1).reverse.join('/')
 
         create_file(create_path, create_name, contents)
-        delete_file(old_path)
+    
+        if(File.dirname(old_path) != create_path)
+          delete_file(old_path)
+        end
       end
 
       def update_file(path, content)
@@ -192,11 +195,11 @@ module ErpTechSvcs
         end
 
         tree_data = {
-            :text => directory.split('/').last,
-            :iconCls => File.directory?(directory) ? 'icon-content' : 'icon-document',
-            :id => directory,
-            :leaf => !File.directory?(directory),
-            :children => []
+          :text => directory.split('/').last,
+          :iconCls => File.directory?(directory) ? 'icon-content' : 'icon-document',
+          :id => directory,
+          :leaf => !File.directory?(directory),
+          :children => []
         }
 
         tree_data[:id].gsub!(root, '') unless options[:keep_full_path]
